Update collections.py
Sequences are covariant in Python. I'm doing this for `Track` but we should probably revisit all the structures.

Also, you're sorting by `start` after checking for `dt` which is probably inefficient. Doing a try/catch would be better, but that only works if `shape.start` errors if you try to access it on a shape that has no `dt`.